requirements.txt
# FloatChat - Ocean Data Query System
# Python Dependencies for Containerized Deployment

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database Connectivity
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2

# AI/ML Integration
cerebras-cloud-sdk==1.4.0
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# API and HTTP
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Date/Time Processing
python-dateutil==2.8.2
pytz==2023.3.post1

# Development and Debugging
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1

# Data Validation
six==1.16.0

# Optional: Caching (if using Redis)
redis==5.0.1

# Optional: Enhanced Logging
structlog==23.2.0

# Optional: Performance Monitoring
psutil==5.9.6

# Optional: Geographic Data Processing
Shapely==2.0.2
Fiona==1.9.5

# Optional: Advanced Data Analysis
scipy==1.11.4
scikit-learn==1.3.2

# Production WSGI Server
gunicorn==21.2.0

# Security
cryptography==41.0.8